Move files into right folders
