+++ client/src/renderer/redux/actions/index.js
@@ -4,3 +4,4 @@ export * from "./cache";
 export * from "./document";
 export * from "./commands";
 export * from "./metadata";
+export * from "./styles";
