+++ client/src/renderer/redux/actions/pageStyles.js
@@ -1,0 +1,7 @@
+/** @format */
+
+export const setPageStyles = (key, data) => ({
+  type: SET_PAGE_STYLES,
+  key,
+  data,
+});
