+++ client/src/renderer/redux/actions/styles.js
@@ -1,0 +1,7 @@
+/** @format */
+
+export const setStyles = (key, data) => ({
+  type: SET_STYLES,
+  key,
+  data,
+});
