+++ client/src/renderer/redux/actions/types.ts
@@ -12,6 +12,10 @@ export const FETCH_DOC_SUCCESS = "FETCH_DOC_SUCCESS";
 // Metadata
 export const SET_METADATA = "SET_METADATA";
 
+//Styles
+export const SET_STYLES = "SET_STYLES";
+export const SET_PAGE_STYLES = "SET_PAGE_STYLES";
+
 // Render caching
 export const CACHE_NODE = "CACHE_NODE";
 export const CACHE_WORKING_AREA = "CACHE_WORKING_AREA";
